requirements.txt
attrs==21.2.0
aws-cdk.assertions==1.132.0
aws-cdk.assets==1.132.0
aws-cdk.aws-acmpca==1.132.0
aws-cdk.aws-apigateway==1.132.0
aws-cdk.aws-applicationautoscaling==1.132.0
aws-cdk.aws-autoscaling==1.132.0
aws-cdk.aws-autoscaling-common==1.132.0
aws-cdk.aws-autoscaling-hooktargets==1.132.0
aws-cdk.aws-certificatemanager==1.132.0
aws-cdk.aws-cloudformation==1.132.0
aws-cdk.aws-cloudfront==1.132.0
aws-cdk.aws-cloudwatch==1.132.0
aws-cdk.aws-codebuild==1.132.0
aws-cdk.aws-codecommit==1.132.0
aws-cdk.aws-codedeploy==1.132.0
aws-cdk.aws-codeguruprofiler==1.132.0
aws-cdk.aws-codepipeline==1.132.0
aws-cdk.aws-codepipeline-actions==1.132.0
aws-cdk.aws-codestarnotifications==1.132.0
aws-cdk.aws-cognito==1.132.0
aws-cdk.aws-ec2==1.132.0
aws-cdk.aws-ecr==1.132.0
aws-cdk.aws-ecr-assets==1.132.0
aws-cdk.aws-ecs==1.132.0
aws-cdk.aws-efs==1.132.0
aws-cdk.aws-elasticloadbalancing==1.132.0
aws-cdk.aws-elasticloadbalancingv2==1.132.0
aws-cdk.aws-events==1.132.0
aws-cdk.aws-events-targets==1.132.0
aws-cdk.aws-globalaccelerator==1.132.0
aws-cdk.aws-iam==1.132.0
aws-cdk.aws-kinesis==1.132.0
aws-cdk.aws-kinesisfirehose==1.132.0
aws-cdk.aws-kms==1.132.0
aws-cdk.aws-lambda==1.132.0
aws-cdk.aws-logs==1.132.0
aws-cdk.aws-route53==1.132.0
aws-cdk.aws-route53-targets==1.132.0
aws-cdk.aws-s3==1.132.0
aws-cdk.aws-s3-assets==1.132.0
aws-cdk.aws-sam==1.132.0
aws-cdk.aws-secretsmanager==1.132.0
aws-cdk.aws-servicediscovery==1.132.0
aws-cdk.aws-signer==1.132.0
aws-cdk.aws-sns==1.132.0
aws-cdk.aws-sns-subscriptions==1.132.0
aws-cdk.aws-sqs==1.132.0
aws-cdk.aws-ssm==1.132.0
aws-cdk.aws-stepfunctions==1.132.0
aws-cdk.cloud-assembly-schema==1.132.0
aws-cdk.core==1.132.0
aws-cdk.custom-resources==1.132.0
aws-cdk.cx-api==1.132.0
aws-cdk.pipelines==1.132.0
aws-cdk.region-info==1.132.0
cattrs==1.8.0
constructs==3.3.161
docopt==0.6.2
dotmap==1.1.16
filelock==2.0.6
futures==3.0.5
Jinja2==2.8
jsii==1.43.0
MarkupSafe==2.0.1
pipelines==0.0.14
publication==0.0.3
python-dateutil==2.8.2
PyYAML==3.11
requests==2.9.1
schema==0.5.0
sh==1.11
six==1.16.0
tornado==4.3
typing-extensions==3.10.0.2
